+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Useful Contacts Page
 - User faculty storage in shared preferences
 - Send bug reports and app crashes details to Sentry
 - Clickable checkbox rows
@@ -32,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the Checkbox background to dark red
 - Changed 'Paragens' to 'Autocarros'
 - Changed Exam's constructor
+- Limited widget dragging to within editing mode
 
 ## [1.1.0] - 2021-04-18